+++ app/controllers/CpanelSliderController.php
@@ -3,19 +3,34 @@
  {
     public function index()
     {
-    	$sliders=Slider::all();
-    	return View::make('Cpanel.slider',compact('sliders'));
+      if(Auth::user()->position=="President"||Auth::user()->position=="Vice President"
+            ||Auth::user()->position=="PR Head"||Auth::user()->position=="PR Member")
+        {
+        	$sliders=Slider::all();
+        	return View::make('Cpanel.slider',compact('sliders'));
+        }
+        else
+          return Redirect::to('cpanel');
     }
     public function create()
     {
-        return View::make('Cpanel.slider-create');
+      if(Auth::user()->position=="President"||Auth::user()->position=="Vice President"
+            ||Auth::user()->position=="PR Head"||Auth::user()->position=="PR Member")
+        {
+
+          return View::make('Cpanel.slider-create');
+        }
+      else
+        return Redirect::to('cpanel');
     }
     public function store()
     {
+      if(Auth::user()->position=="President"||Auth::user()->position=="Vice President"
+            ||Auth::user()->position=="PR Head"||Auth::user()->position=="PR Member")
+      {
     	$validate=Validator::make(Input::all(),array(
        'name'=>'required',
        'image'=>'required',
-       'link'=>'required',
        'description'=>'required',
       ));
 		 if($validate->fails())
@@ -23,32 +38,36 @@
 		 else
 		 {
 
-		 	$slider=new Slider();    
+		 	      $slider=new Slider();    
             $slider->name=Input::get('name');
             $slider->updated_by=Auth::user()->username;
-		    $name=Input::file('image')->getClientOriginalName();
-		    Input::file('image')->move('images/sliders',$name);
-		    $slider->imgPath='images/sliders/' . $name;
-		 	  $slider->link=Input::get('link');
-		 	  $slider->description=Input::get('description');
-		  	if(Input::get('availibility')=='on')
-		 	    $slider->availibility=true;
-		    else
-		 	    $slider->availibility=false;
+            
+		      $name=$slider->name . '-' . time() . '.' .Input::file('image')->getClientOriginalName();
+		      Input::file('image')->move('images/sliders',$name);
+		      $slider->imgPath='images/sliders/' . $name;
+		 	    $slider->description=Input::get('description');
+		  	  if(Input::get('availibility')=='on')
+		 	       $slider->availibility=true;
+		      else
+		 	       $slider->availibility=false;
 		  }
 
 		 if($slider->save())
         	return Redirect::to('cpanel/sliders');
         else
         	return "an fatel error has been occuered !";
-        
+      }
+      else
+        return Redirect::to('cpanel');
 
     }
     public function update($id)
     {
+     if(Auth::user()->position=="President"||Auth::user()->position=="Vice President"
+            ||Auth::user()->position=="PR Head"||Auth::user()->position=="PR Member")
+      {
     	$validate=Validator::make(Input::all(),array(
        'name'=>'required',
-       'link'=>'required',
        'description'=>'required',
       ));
 		 if($validate->fails())
@@ -57,7 +76,6 @@
 		 {
 		 	$slider = Slider::find($id);
 		 	$slider->name=Input::get('name');
-		 	$slider->link=Input::get('link');
 		 	$slider->updated_by=Auth::user()->username;
 		 	$slider->description=Input::get('description');
 		 	if(Input::get('availibility')=='on')
@@ -67,7 +85,7 @@
 		    if(Input::hasFile('image'))
 		    {
 		      unlink(public_path() . $slider->imaPath);
-		      $name=Input::file('image')->getClientOriginalName();
+		      $name=$slider->name . '-' . time() . '.' .Input::file('image')->getClientOriginalName();
 		      Input::file('image')->move('images/sliders',$name);
 		      $slider->imgPath='images/sliders/' . $name;
 		    }
@@ -77,20 +95,35 @@
         	return Redirect::to('cpanel/sliders');
         else
         	return "an fatel error has been occured";
+      }
+      else
+        return Redirect::to('cpanel');
     }
     public function edit($id)
     {
-    	$slider = Slider::find($id);		 
-		return View::make('Cpanel.slider-edit',compact('slider'));
+       if(Auth::user()->position=="President"||Auth::user()->position=="Vice President"
+            ||Auth::user()->position=="PR Head"||Auth::user()->position=="PR Member")
+       {
+    	    $slider = Slider::find($id);		 
+		      return View::make('Cpanel.slider-edit',compact('slider'));
+       }
+      else
+       return Redirect::to('cpanel');
     }
     public function destroy($id)
     {
-      $deletedrow = Slider::find($id);
-      unlink(public_path() . $deletedrow->imaPath);
-      $deletedrow->delete();
-		    if($deletedrow)
-		    	return Redirect::to('cpanel/sliders');
-		    else
-			    return "an error has been occured" ;
+       if(Auth::user()->position=="President"||Auth::user()->position=="Vice President"
+            ||Auth::user()->position=="PR Head"||Auth::user()->position=="PR Member")
+       {
+          $deletedrow = Slider::find($id);
+          unlink(public_path() . $deletedrow->imaPath);
+          $deletedrow->delete();
+		      if($deletedrow)
+		    	   return Redirect::to('cpanel/sliders');
+		      else
+			       return "an error has been occured" ;
+      }
+      else
+        return Redirect::to('cpanel');
     }
  }
